requirements.txt
paho_mqtt==1.6.1
python-dotenv==1.0.0
Requests==2.31.0
